requirements.txt
black
pyright
requests
python-dotenv
rich
